pyproject.toml
[build-system]
requires = ["setuptools >= 65.5.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
requires-python = ">=3.11,<4"
name = "semkon"
dependencies = [
    "claude-agent-sdk==0.1.13",
    "cyclopts==4.3.0",
    "gitignore-parser==0.1.13",
    "pydantic==2.12.5",
    "pydantic-ai==1.27.0",
    "PyYAML==6.0.3",
    "tqdm==4.67.1",
]
version = "2.0.0"

[project.optional-dependencies]
dev = [
    "ruff",
    "ty",
]

[project.scripts]
semkon = "semkon.main:app"

[tool.setuptools.packages.find]
include = ["semkon"]
namespaces = false

[tool.ruff]
lint.select = ["F401"]
line-length = 80
